splitting and ordering api tests into multiple files
